+++ src/crawlee/basic_crawler/_basic_crawler.py
@@ -473,6 +473,8 @@ class BasicCrawler(Generic[TCrawlingContext]):
         self._running = False
         self._has_finished_before = True
 
+        await self._save_crawler_state()
+
         final_statistics = self._statistics.calculate()
         self._logger.info(f'Final request statistics:\n{final_statistics.to_table()}')
 
@@ -526,6 +528,16 @@ class BasicCrawler(Generic[TCrawlingContext]):
             wait_for_all_requests_to_be_added=wait_for_all_requests_to_be_added,
             wait_for_all_requests_to_be_added_timeout=wait_for_all_requests_to_be_added_timeout,
         )
+
+    async def _use_state(
+        self, key: str, default_value: dict[str, JsonSerializable] | None = None
+    ) -> dict[str, JsonSerializable]:
+        store = await self.get_key_value_store()
+        return await store.get_auto_saved_value(key, default_value)
+
+    async def _save_crawler_state(self) -> None:
+        store = await self.get_key_value_store()
+        await store.persist_autosaved_values()
 
     async def get_data(
         self,
@@ -948,6 +960,7 @@ class BasicCrawler(Generic[TCrawlingContext]):
             add_requests=result.add_requests,
             push_data=result.push_data,
             get_key_value_store=result.get_key_value_store,
+            use_state=self._use_state,
             log=self._logger,
         )