+++ tests/unit/basic_crawler/test_basic_crawler.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 import asyncio
 import json
 import logging
+import os
 from collections import Counter
 from dataclasses import dataclass
 from datetime import timedelta
@@ -24,10 +25,19 @@ from crawlee.statistics import FinalStatistics
 from crawlee.storages import Dataset, KeyValueStore, RequestList, RequestQueue, RequestSourceTandem
 
 if TYPE_CHECKING:
-    from collections.abc import Sequence
+    from collections.abc import AsyncGenerator, Sequence
 
     import respx
     from yarl import URL
+
+    from crawlee._types import JsonSerializable
+
+
+@pytest.fixture
+async def key_value_store() -> AsyncGenerator[KeyValueStore, None]:
+    kvs = await KeyValueStore.open()
+    yield kvs
+    await kvs.drop()
 
 
 async def test_processes_requests_from_explicit_queue() -> None:
@@ -675,6 +685,63 @@ async def test_context_update_kv_store() -> None:
     assert (await store.get_value('foo')) == 'bar'
 
 
+async def test_context_use_state(key_value_store: KeyValueStore) -> None:
+    crawler = BasicCrawler()
+
+    @crawler.router.default_handler
+    async def handler(context: BasicCrawlingContext) -> None:
+        await context.use_state('state', {'hello': 'world'})
+
+    await crawler.run(['https://hello.world'])
+
+    store = await crawler.get_key_value_store()
+
+    assert (await store.get_value('state')) == {'hello': 'world'}
+
+
+async def test_context_handlers_use_state(key_value_store: KeyValueStore) -> None:
+    state_in_handler_one: dict[str, JsonSerializable] = {}
+    state_in_handler_two: dict[str, JsonSerializable] = {}
+    state_in_handler_three: dict[str, JsonSerializable] = {}
+
+    crawler = BasicCrawler()
+
+    @crawler.router.handler('one')
+    async def handler_one(context: BasicCrawlingContext) -> None:
+        state = await context.use_state('state', {'hello': 'world'})
+        state_in_handler_one.update(state)
+        state['hello'] = 'new_world'
+        await context.add_requests([Request.from_url('https://crawlee.dev/docs/quick-start', label='two')])
+
+    @crawler.router.handler('two')
+    async def handler_two(context: BasicCrawlingContext) -> None:
+        state = await context.use_state('state', {'hello': 'world'})
+        state_in_handler_two.update(state)
+        state['hello'] = 'last_world'
+
+    @crawler.router.handler('three')
+    async def handler_three(context: BasicCrawlingContext) -> None:
+        state = await context.use_state('state', {'hello': 'world'})
+        state_in_handler_three.update(state)
+
+    await crawler.run([Request.from_url('https://crawlee.dev/', label='one')])
+    await crawler.run([Request.from_url('https://crawlee.dev/docs/examples', label='three')])
+
+    # The state in handler_one must match the default state
+    assert state_in_handler_one == {'hello': 'world'}
+
+    # The state in handler_two must match the state updated in handler_one
+    assert state_in_handler_two == {'hello': 'new_world'}
+
+    # The state in handler_three must match the final state updated in previous run
+    assert state_in_handler_three == {'hello': 'last_world'}
+
+    store = await crawler.get_key_value_store()
+
+    # The state in the KVS must match with the last set state
+    assert (await store.get_value('state')) == {'hello': 'last_world'}
+
+
 async def test_max_requests_per_crawl(httpbin: URL) -> None:
     start_urls = [
         str(httpbin / '1'),
@@ -821,6 +888,7 @@ async def test_respects_no_persist_storage() -> None:
     assert not datasets_path.exists() or list(datasets_path.iterdir()) == []
 
 
+@pytest.mark.skipif(os.name == 'nt' and 'CI' in os.environ, reason='Skipped in Windows CI')
 async def test_logs_final_statistics(monkeypatch: pytest.MonkeyPatch, caplog: pytest.LogCaptureFixture) -> None:
     # Set the log level to INFO to capture the final statistics log.
     caplog.set_level(logging.INFO)